requirements.txt
fastapi
uvicorn
websockets
pydantic[email]
httpx
motor
passlib[bcrypt]
python-jose[cryptography]
bcrypt
